requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.2
django-model-utils==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-extensions==0.7.1
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.16.0
psycopg2==2.9.3
PyJWT==2.5.0
pyrsistent==0.18.1
pytz==2022.4
PyYAML==6.0
rest-framework-roles==0.4.1
sqlparse==0.4.3
tzdata==2022.4
uritemplate==4.1.1
